requirements.txt
Django>=1.8,<1.9
django-pagedown
django-markdown-deux
less
